requirements.txt
praw==7.7.1
pandas==2.2.0
pytz==2024.1
tqdm==4.66.1
datasets==2.18.0
huggingface-hub==0.20.3
pyyaml==6.0.1 
python-dotenv>=1.0.0
